+++ embedded_c_sdk_bind_hal/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.4.0 - 2024-08-26
+
+- replace print-log-ufmt featue by print-log-csdk
+- use `embedded-c-sdk-bind-print-macros` crate print/println macro to format print
+
 ## 0.3.4 - 2024-08-22
 
 - add print-log-ufmt featue to optimize fmt code size